requirements.txt
openai
pyyaml
tenacity
tiktoken
